Add get_queryset method to expose QuerySet objects
Downstream projects desire the ability to retrieve the individual
`QuerySet` objects which form the `QuerySetSequence`.  This commit adds
a new `get_querysets` method as part of a stable public API to return
the list of `QuerySet` objects that the sequence object is currently
wrapping.

Tests and documentation updates are present, noting some of the caveats
around this method. Notably, the component querysets will differ from
those used to construct the original `QuerySetSequence` object if
methods have been invoked which modify the sequence. Also that the
order in which the `QuerySet` objects are returned is not guaranteed
to be constant.

Fixes #21